==> CHECK_CEDX/3_make_variable.ipynb <==
# Import libs
import os
import pandas as pd
pd.options.mode.chained_assignment = None  # default='warn'
import numpy as np
from datetime import datetime
import importlib
import global_settings
importlib.reload(global_settings)

setts = global_settings.Settings()

root_folder = setts.root_folder

input_tmp_folder = setts.tmp1_folder
output_folder = setts.tmp2_folder


# DCSI

### データの入力
INPUT_PATH_DCSI = input_tmp_folder + "/DCSI.csv"
df = pd.read_csv(INPUT_PATH_DCSI, encoding='utf-8')

### カラムの抽出
collist = [
"UID",
"Month",
"Year",
"Dealer",
"Region",
"REMINDER_",
"RECEPTION",
"CUSTOMER_LOUNGE",
"DELIVERY",
"REPAIR_QUALITY",
"FACILITY",
"Q30Y",
"OVERALL_SATISFACTION",
"Free_comment"
]

df = df[collist]

### データの出力
OUTPUT_PATH_DCSI = output_folder + "/DCSI_ver1.csv"
df.to_csv(OUTPUT_PATH_DCSI, encoding='utf-8', index=False)

# DCSIに関する追加の処理

# データの入力
df = pd.read_csv(INPUT_PATH_DCSI, encoding='utf-8')

# カラムの抽出・データ型の変換
df_DCSI = df[['UID', 'Q35Y']].copy()
df_DCSI['follow_call'] = pd.to_numeric(df_DCSI['Q35Y'], errors='coerce')

# follow_callごとの処理
# Processing each follow call
df_DCSI.fillna({'follow_call': 0}, inplace=True)
df_DCSI_output = df_DCSI.groupby(by='UID').agg(
    follow_call=('follow_call', 'max')
).reset_index()

# ↑のセルで作成したデータの入力
# File created by cell above
df_DCSI_new_name = pd.read_csv(OUTPUT_PATH_DCSI)

# データの結合
df_DCSI_ver2 = pd.merge(df_DCSI_new_name, df_DCSI_output, on = 'UID', how = 'left')

# データの出力
df_DCSI_ver2.to_csv(output_folder + "/DCSI_output_ver2.csv", encoding='utf-8', index=False)

# My Honda

### データの入力
INPUT_PATH_MH = input_tmp_folder + "/MH.csv"
df = pd.read_csv(INPUT_PATH_MH, encoding='utf-8')

### カラムの抽出
usecols = [
    'rate_service_point',
    'rate_mechanic_point',
    'rate_chatbot_point',
    'rate_chatbot_agent_point',
    'UID',
]

df_MH = df[usecols]


### 最後の点数を抽出
# rate_service_point, rate_mechanic_pointについて、小数表記のため3文字分を抽出
# rate_service_point
rate_service_point_5str_1 = df_MH.loc[:, 'rate_service_point'].str[-5:-1] 
rate_service_point_5str_2 = rate_service_point_5str_1.str[-3:]
# rate_mechanic_point
rate_mechanic_point_5str_1 = df_MH.loc[:, 'rate_mechanic_point'].str[-5:-1] 
rate_mechanic_point_5str_2 = rate_mechanic_point_5str_1.str[-3:]
# rate_chatbot_point, rate_chatbot_pointについて、整数表記のため1文字分を抽出
# rate_chatbot_point
rate_chatbot_point_5str_1 = df_MH.loc[:, 'rate_chatbot_point'].str[-2:-1] 
rate_chatbot_point_5str_2 = rate_chatbot_point_5str_1.str[-1:]
# rate_chatbot_agent_point
rate_chatbot_agent_point_5str_1 = df_MH.loc[:, 'rate_chatbot_agent_point'].str[-5:-1] 
rate_chatbot_agent_point_5str_2 = rate_chatbot_agent_point_5str_1.str[-1:]


### 文字列から数値への型変換
# rate_service_point
df_rate_service_point_num = pd.to_numeric(rate_service_point_5str_2, errors='coerce') 
# rate_mechanic_point
df_rate_mechanic_point_num =  pd.to_numeric(rate_mechanic_point_5str_2, errors='coerce')
# rate_chatbot_point
df_rate_chatbot_point_num =  pd.to_numeric(rate_chatbot_point_5str_2, errors='coerce')
# rate_chatbot_agent_point
df_rate_chatbot_agent_point_num =  pd.to_numeric(rate_chatbot_agent_point_5str_2, errors='coerce')
### dataframeへの変換
df_rate_service_point_num = pd.DataFrame(df_rate_service_point_num)
df_rate_mechanic_point_num =  pd.DataFrame(df_rate_mechanic_point_num)
df_rate_chatbot_point_num =  pd.DataFrame(df_rate_chatbot_point_num)
df_rate_chatbot_agent_point_num =  pd.DataFrame(df_rate_chatbot_agent_point_num)


### - カラム名の変更とカラム抽出
# - rate_service_point
df_rate_service_point_num.rename(columns={
    'rate_service_point' : 'rate_service_point_NUM'
}, inplace=True)
# - rate_mechanic_point
df_rate_mechanic_point_num.rename(columns={
    'rate_mechanic_point' : 'rate_mechanic_point_NUM'
}, inplace=True)
# - rate_chatbot_point
df_rate_chatbot_point_num.rename(columns={
    'rate_chatbot_point' : 'rate_chatbot_point_NUM'
}, inplace=True)
# - rate_chatbot_agent_point
df_rate_chatbot_agent_point_num.rename(columns={
    'rate_chatbot_agent_point': 'rate_chatbot_agent_point_NUM'
}, inplace=True)
### データの結合
df_MH2 = pd.concat([df_MH, df_rate_service_point_num, df_rate_mechanic_point_num, df_rate_chatbot_point_num, df_rate_chatbot_agent_point_num], axis = 'columns')
### 必要カラムの抽出
df_MH2.drop(columns=['rate_service_point', 'rate_mechanic_point', 'rate_chatbot_point', 'rate_chatbot_agent_point'], inplace=True)
# カラム名の編集
df_MH2.rename(columns={
    'rate_service_point_NUM' : 'rate_service_point',
    'rate_mechanic_point_NUM' : 'rate_mechanic_point',
    'rate_chatbot_point_NUM' : 'rate_chatbot_point',
    'rate_chatbot_agent_point_NUM' : 'rate_chatbot_agent_point',
}, inplace=True)
# 必要カラムの抽出
df_MH3 = df_MH2[['UID',  'rate_service_point', 'rate_mechanic_point', 'rate_chatbot_point', 'rate_chatbot_agent_point']]

### uniqueなUIDごとに計算する
# uniqueなUID（昇順）の取得
df_uid = df_MH3['UID']
uid_unique = df_uid.unique()
df_uid_unique = pd.DataFrame({'UID': uid_unique})

# UIDの出現回数をカウント
df_uid_count = df_uid.value_counts().rename_axis('UID').reset_index(name='UID_counts_MyHonda')

# uniqueなuid（昇順）に出現回数を付与
df_uid_unique_count = pd.merge(df_uid_unique, df_uid_count, on='UID', how='left')

# UIDごとに、各pointの平均をとる
df_MyHonda = df_MH2.groupby(['UID'], dropna=False, as_index=False).agg(
        rate_service_point = ('rate_service_point', 'mean'),
        rate_mechanic_point = ('rate_mechanic_point', 'mean'),
        rate_chatbot_point = ('rate_chatbot_point', 'mean'),
        rate_chatbot_agent_point = ('rate_chatbot_agent_point', 'mean'),
)
# uniqueなUIDに結合
df_MyHonda = pd.merge(df_uid_unique_count, df_MyHonda, on='UID', how='left')

# MyhondaID有無を表すカラム：UIIDがあればMyHondaIDありとする
# Check to see if MyHondaID exist or not, if UID exist then MyHondaID exist
df_MyHonda.loc[:, ["MyHonda有無"]] = '1'

### データの出力
OUTPUT_PATH_MH = output_folder + "/MH_new_name.csv"
df_MyHonda.to_csv(OUTPUT_PATH_MH, encoding='utf-8', index=False)

### データの入力
INPUT_PATH_MH = input_tmp_folder + "/MH.csv"
df = pd.read_csv(INPUT_PATH_MH, encoding='utf-8')

### カラムの抽出
usecols = [
    'rate_service_point',
    'rate_mechanic_point',
    'rate_chatbot_point',
    'rate_chatbot_agent_point',
    'UID',
]

df_MH = df[usecols]

### 最後の点数を抽出
# rate_service_point, rate_mechanic_pointについて、小数表記のため3文字分を抽出
# rate_service_point
rate_service_point_5str_1 = df_MH.loc[:, 'rate_service_point'].astype(str).str[:]
rate_service_point_5str_2 = rate_service_point_5str_1.str[:]
# rate_mechanic_point
rate_mechanic_point_5str_1 = df_MH.loc[:, 'rate_mechanic_point'].astype(str).str[:]
rate_mechanic_point_5str_2 = rate_mechanic_point_5str_1.str[:]
# rate_chatbot_point, rate_chatbot_pointについて、整数表記のため1文字分を抽出
# rate_chatbot_point
rate_chatbot_point_5str_1 = df_MH.loc[:, 'rate_chatbot_point'].astype(str).str[:]
rate_chatbot_point_5str_2 = rate_chatbot_point_5str_1.str[:]
# # rate_chatbot_agent_point
rate_chatbot_agent_point_5str_1 = df_MH.loc[:, 'rate_chatbot_agent_point'].astype(str).str[:]
rate_chatbot_agent_point_5str_2 = rate_chatbot_agent_point_5str_1.str[:]


### 文字列から数値への型変換
# rate_service_point
df_rate_service_point_num = pd.to_numeric(rate_service_point_5str_2, errors='coerce')
# rate_mechanic_point
df_rate_mechanic_point_num =  pd.to_numeric(rate_mechanic_point_5str_2, errors='coerce')
# rate_chatbot_point
df_rate_chatbot_point_num =  pd.to_numeric(rate_chatbot_point_5str_2, errors='coerce')
# rate_chatbot_agent_point
df_rate_chatbot_agent_point_num =  pd.to_numeric(rate_chatbot_agent_point_5str_2, errors='coerce')
### dataframeへの変換

df_rate_service_point_num = pd.DataFrame(df_rate_service_point_num)
df_rate_mechanic_point_num =  pd.DataFrame(df_rate_mechanic_point_num)
df_rate_chatbot_point_num =  pd.DataFrame(df_rate_chatbot_point_num)
df_rate_chatbot_agent_point_num =  pd.DataFrame(df_rate_chatbot_agent_point_num)


### - カラム名の変更とカラム抽出
# - rate_service_point
df_rate_service_point_num.rename(columns={
    'rate_service_point' : 'rate_service_point_NUM'
}, inplace=True)
# - rate_mechanic_point
df_rate_mechanic_point_num.rename(columns={
    'rate_mechanic_point' : 'rate_mechanic_point_NUM'
}, inplace=True)
# - rate_chatbot_point
df_rate_chatbot_point_num.rename(columns={
    'rate_chatbot_point' : 'rate_chatbot_point_NUM'
}, inplace=True)
# # - rate_chatbot_agent_point
df_rate_chatbot_agent_point_num.rename(columns={
    'rate_chatbot_agent_point': 'rate_chatbot_agent_point_NUM'
}, inplace=True)
### データの結合
#df_MH2 = pd.concat([df_MH, df_rate_service_point_num, df_rate_mechanic_point_num])#, df_rate_chatbot_point_num, df_rate_chatbot_agent_point_num], axis = 'columns')
# df_MH2 = pd.concat([df_MH, df_rate_service_point_num, df_rate_mechanic_point_num], ignore_index=True)#, df_rate_chatbot_point_num, df_rate_chatbot_agent_point_num], axis = 'columns')
df_MH2 = df_MH.copy()
df_MH2['rate_service_point'] = df_rate_service_point_num['rate_service_point_NUM']
df_MH2['rate_mechanic_point'] = df_rate_mechanic_point_num['rate_mechanic_point_NUM']
df_MH2['rate_chatbot_point'] = df_rate_chatbot_point_num['rate_chatbot_point_NUM']
df_MH2['rate_chatbot_agent_point'] = df_rate_chatbot_agent_point_num['rate_chatbot_agent_point_NUM']


df_MH3 = df_MH2[['UID',  'rate_service_point', 'rate_mechanic_point','rate_chatbot_point', 'rate_chatbot_agent_point']]

### uniqueなUIDごとに計算する
# uniqueなUID（昇順）の取得
df_uid = df_MH3['UID'].dropna()
uid_unique = df_uid.unique()
df_uid_unique = pd.DataFrame({'UID': uid_unique})

# UIDの出現回数をカウント
df_uid_count = df_uid.value_counts().rename_axis('UID').reset_index(name='UID_counts_MyHonda')

# uniqueなuid（昇順）に出現回数を付与
df_uid_unique_count = pd.merge(df_uid_unique, df_uid_count, on='UID', how='left')
# UIDごとに、各pointの平均をとる

grouped_service_point = df_MH2.groupby(['UID'])['rate_service_point'].mean().reset_index()
grouped_mechanic_point = df_MH2.groupby(['UID'])['rate_mechanic_point'].mean().reset_index()
grouped_chatbot_point = df_MH2.groupby(['UID'])['rate_chatbot_point'].mean().reset_index()
grouped_agent_point = df_MH2.groupby(['UID'])['rate_chatbot_agent_point'].mean().reset_index()

df_MyHonda = pd.merge(grouped_service_point, grouped_mechanic_point, on='UID', how='left')
df_MyHonda = pd.merge(df_MyHonda, grouped_chatbot_point, on='UID', how='left')
df_MyHonda = pd.merge(df_MyHonda, grouped_agent_point, on='UID', how='left')

df_MyHonda = pd.merge(df_uid_unique_count, df_MyHonda, on='UID', how='left')
df_MyHonda.drop_duplicates(subset='UID', inplace=True)

# MyhondaID有無を表すカラム：UIIDがあればMyHondaIDありとする
# Check to see if MyHondaID exist or not, if UID exist then MyHondaID exist
df_MyHonda.loc[:, ["MyHonda有無"]] = '1'

### データの出力
OUTPUT_PATH_MH = output_folder + "/MH_new_name.csv"
df_MyHonda.to_csv(OUTPUT_PATH_MH, encoding='utf-8', index=False)

# booking_service

# UIDから出現回数と有無を表すフラグを立てる関数
def uid_unique_count_flag(df, filename):
    # df_uid: UIDのみのタプル
    df_uid = df['UID']

    # uniqueなUIDの取得
    uid_unique = df_uid.unique()
    df_uid_unique = pd.DataFrame({'UID': uid_unique})
    
    # UIDの出現回数
    column_name = 'UID_counts_' + filename
    df_uid_count = df_uid.value_counts().rename_axis('UID').reset_index(name=column_name)
    
    # uniqueなuidと出現回数のカラムの結合
    df_uid_unique_count = pd.merge(df_uid_unique, df_uid_count, on='UID', how='left')
    
    # UID有無を表すカラム
    df_uid_unique_count.loc[:, ["UID有無"]] = '1'

    return df_uid_unique_count

### データの入力
# BOOK data's field has a space before most columns
INPUT_PATH_BOOK = input_tmp_folder + "/book.csv"
df = pd.read_csv(INPUT_PATH_BOOK, encoding='utf-8')
#df_book = df.iloc[:, [-1, 1, 3, 5, 13]]
df_book = df[['UID', 'dealer_code', 'job_service_type', 'emp_id', 'booking_date']]

df_book_uid = uid_unique_count_flag(df_book, 'book')

### emp_idについての処理
df_book["emp_id_num"] = pd.to_numeric(df_book['emp_id'], errors='coerce') # <==== emp_id has a space before it
df_book.loc[~df_book['emp_id_num'].isna(), ['emp_choice']] = 1
df_book.loc[df_book['emp_id_num'].isna(), ['emp_choice']] = 0
df_emp = df_book.groupby(by='UID').agg(
    emp_id_count=('emp_choice', 'sum'),
    emp_id_有無=('emp_choice', 'max'),
).reset_index()

### booking_dateについての処理
df_booking_date = df_book.groupby(by="UID").aggregate(
    booking_date_count = ('booking_date', 'nunique'), # <==== booking_date has a space before it
).reset_index()


### 結合
df_book_uid = pd.merge(df_book_uid, df_emp, on = 'UID', how='left')
df_book_uid = pd.merge(df_book_uid, df_booking_date, on = 'UID', how='left')

### データの出力
OUTPUT_PATH_book = output_folder + "/book_new_name.csv"
df_book_uid.to_csv(OUTPUT_PATH_book, encoding='utf-8', index=False)

# notificiation_tracking, event_tracking
- 処理ファイル一覧
    - event_v_promo.csv
    - notif_vpi
    - notif_vpm
    - notif_v_pro
    - notif_v_spa
    - notif_book_pi
    - notif_book_pm
    - notif_book_spa

# UIDから出現回数と有無を表すフラグを立てる関数
def uid_unique_count_flag(df, filename):
    # df_uid: UIDのタプル
    df_uid = df['UID']

    # uniqueなUIDの取得
    uid_unique = df_uid.unique()
    df_uid_unique = pd.DataFrame({'UID': uid_unique})

    # UIDの出現回数をカウント
    column_name = 'UID_counts'
    df_uid_count = df_uid.value_counts().rename_axis('UID').reset_index(name=column_name)
    
    # uniqueなuidと出現回数のカラムの結合
    df_uid_unique_count = pd.merge(df_uid_unique, df_uid_count, on='UID', how='left')
    

    # UID有無を表すカラム
    df_uid_unique_count[filename] = '1'

    if filename == "event_v_promo":
        # event_v_promoの場合
        #   UID有無を表すカラムのカラム名をcoupon_useに変更
        df_uid_unique_count = df_uid_unique_count.rename(columns={filename: "coupon_use"})
        #   UIDの出現回数を表すカラムのカラム名をevent_v_promoに変更
        df_uid_unique_count = df_uid_unique_count.rename(columns={"UID_counts": filename})

    elif filename == "notif_v_pro":
        # notif_v_proの場合
        #   UID有無を表すカラムを削除
        df_uid_unique_count = df_uid_unique_count.drop(columns=filename, errors='ignore')
        #   UIDの出現回数を表すカラムのカラム名をevent_v_promoに変更
        df_uid_unique_count = df_uid_unique_count.rename(columns={"UID_counts": "coupon_view"})

    else:
        # その他の場合
        #   UIDの出現回数を表すカラムを削除
        df_uid_unique_count = df_uid_unique_count.drop(columns="UID_counts")

    return df_uid_unique_count

### 入力パス、出力パスの指定
PATH_DIR = input_tmp_folder
OUTPUT_PATH_DIR = output_folder

input_filename_list = [
    "event_v_promo", #  view_promotion_notification
    "notif_vpi", # view_pi_remind_notification
    "notif_vpm", # view_pm_remind_notification
    "notif_v_pro", # view_promotion_notification
    "notif_v_spa", # view_sparepart_notification
    "notif_book_pi", # booking_pi_remind_notification
    "notif_book_pm", # booking_pm_service
    "notif_book_spa", # booking_sparepart
]

input_path_list = []
output_path_list = []

for name in input_filename_list:
    # input
    name_csv = name + '.csv'
    input_path = os.path.join(PATH_DIR, name_csv)
    input_path_list.append(input_path)

    # output
    output_name = name_csv.replace(".csv", "_new_name.csv")
    output_path = os.path.join(OUTPUT_PATH_DIR, output_name)
    output_path_list.append(output_path)

### 各ファイルについて処理
for input, output, filename in zip(input_path_list, output_path_list, input_filename_list):
    df_input = pd.read_csv(input, encoding="utf-8")
    df_output = uid_unique_count_flag(df_input, filename)
    df_output.to_csv(output, encoding="utf-8", index=False)


### notif_v_pro, event_v_promoについて、個別に対応
df_notif = "↑のセルで生成したファイルのパスを個別に指定"
df_event = "↑のセルで生成したファイルのパスを個別に指定"

# SCA

### データの入力
INPUT_PATH_SCA = input_tmp_folder + "/SCA.csv"
df = pd.read_csv(INPUT_PATH_SCA, encoding='utf-8')
df['start_date_df'] = pd.to_datetime(df['start_date'])

### カラムの抽出
df_sca_input = df[['UID', 'start_date_df']]

### UIDごと
df_sca = df_sca_input.groupby(by='UID').agg(
    sca_visit_num=('start_date_df', 'count'),
).reset_index()

### 出力
OUTPUT_PATH_SCA = output_folder + "/SCA_output.csv"
df_sca.to_csv(OUTPUT_PATH_SCA, encoding='utf-8', index=False)

# Comp

### データの入力
INPUT_PATH_COMP = input_tmp_folder + "/comp.csv"
df = pd.read_csv(INPUT_PATH_COMP, encoding='utf-8')
#df['start_date_df'] = pd.to_datetime(df['start_date'])
df['start_date_df'] = pd.to_datetime(df['Date/Time Opened'])


### カラムの抽出
df_comp_input = df.loc[:, ['UID', 'CR Code Full']]
df_comp_input['suggestion'] = 0


# CR code full（提案）に属するフラグを立てる
list_suggestion = ["CA99", " CB98", " CB99", " CC99", " CD99", " CE99", " CF99", " CZ99", " CZ80", " CZ81"]
df_comp_input.loc[df_comp_input["CR Code Full"].isin(list_suggestion), ['suggestion']] = 1 


### UIDごとの提案回数をカウント
df_comp = df_comp_input.groupby(by='UID').agg(
    CR_feedback=('suggestion', 'count'),
).reset_index()
df_comp['CR'] = '1'

### 出力
OUTPUT_PATH_COMP = output_folder + "/comp_output.csv"
df_comp.to_csv(OUTPUT_PATH_COMP, encoding='utf-8', index=False)


# ew

### データの入力
INPUT_PATH_EW = input_tmp_folder + "/ew.csv"
df = pd.read_csv(INPUT_PATH_EW, encoding='utf-8')
#df['start_date_df'] = pd.to_datetime(df['start_date'])
df['start_date_df'] = pd.to_datetime(df['EW START '])

### カラムの抽出・カラム名の修正
df_ew = df.copy()
# カラム名の修正
df_ew.rename(columns={
    'Lo<U+1EA1>i goi' : 'Package_Type',
    'Loại gói' : 'Package_Type',
}, inplace=True)
df_ew = df_ew[['start_date_df', 'Package_Type', 'NET', 'UID']]


### UIDごとの算出
df_ew_NewTotal = df_ew.groupby(by='UID').agg(
    ew_sales=('NET', 'sum'),
).reset_index()

### ALL_carフラグを立てて算出
df_ew['ALL_car'] = 0
df_ew.loc[df_ew["Package_Type"]=='All car', ['ALL_car']] = 1 
df_ew_Allcar = df_ew.groupby(by='UID').agg(
    ew_all_car = ('ALL_car', 'max'),
    ew_nun = ('ALL_car', 'sum')
)

### 結合
df_ew_output = pd.merge(df_ew_NewTotal, df_ew_Allcar, on='UID', how = 'outer')

### 出力
OUTPUT_PATH = output_folder + "/ew_output.csv"
df_ew_output.to_csv(OUTPUT_PATH, encoding='utf-8', index=False)

# chat

# データの入力
INPUT_PATH = input_tmp_folder + "/chat.csv"
#collist = [10, 5, 6]

collist = ['UID', 'Business', 'Field']
df = pd.read_csv(INPUT_PATH, usecols=collist, encoding='utf-8')

### UIDの出現回数のカウント
df_uid_count = df['UID'].value_counts().rename_axis('UID').reset_index(name='chat_num')

### Service Partについて
# Business=AMかつField=Servideのレコードにフラグ1、その他を0として→agg max
df.loc[(df['Business'] == 'AM') & (df['Field'] == 'Service'), ['Business_Field']] = 1
df.loc[~((df['Business'] == 'AM') & (df['Field'] == 'Service')), ['Business_Field']] = 0
df_service_part = df.groupby(by='UID').agg(
    service_part = ('Business_Field', 'max')
).reset_index()

### 結合
df_chat = pd.merge(df_uid_count, df_service_part, on = 'UID', how='left')

# UID有無を表す列を作成

#df_chat.loc[:, 'chat'] = 1
df_chat['chat'] = 1


# 必要カラムの抽出
df_chat_output = df_chat[['UID', 'chat','chat_num' , 'service_part']]

# データの出力
OUTPUT_PATH = output_folder + "/chat_output.csv"
df_chat.to_csv(OUTPUT_PATH, index=False, encoding=False)

# test_drive
#PATH_DIR = "C:/Users/J0135690/OneDrive - Honda/ドキュメント/プロジェクト/CeDX/データ"
#INPUT_PATH = os.path.join(PATH_DIR, "test_drive.csv")
#OUTPUT_PATH = os.path.join(PATH_DIR, "test_drive_output.csv")

# test_drive

### データの入力
INPUT_PATH = input_tmp_folder + "/test_drive.csv"
df = pd.read_csv(INPUT_PATH, encoding='utf-8')

# UIDごとにevent_date回数をカウント
df_test_drive = df.groupby(by='UID').agg(
    test_drive = ('event_date', 'count')
).reset_index()

### データの出力
OUTPUT_PATH = output_folder + "/test_drive_output.csv"
df_test_drive.to_csv(OUTPUT_PATH, encoding='utf-8', index=False)

# view_test_drive

### データの入力
INPUT_PATH = input_tmp_folder + "/view_test_drive.csv"
df = pd.read_csv(INPUT_PATH, encoding='utf-8')

# UIDごとにevent_date回数をカウント
df_view_test_drive = df.groupby(by='UID').agg(
    view_test_drive=('event_date', 'count')
).reset_index()

### データの出力
OUTPUT_PATH = output_folder + "/view_test_drive_output.csv"
df_view_test_drive.to_csv(OUTPUT_PATH, encoding='utf-8', index=False)

# view_promotion_notification

# データの入力
INPUT_PATH = input_tmp_folder + "/view_promotion_notification.csv"
df = pd.read_csv(INPUT_PATH, encoding='utf-8')

# UIDごとにevent_date回数をカウント
df_view_promotion_notification = df.groupby(by='UID').agg(
    view_promotion_notification=('event_date', 'count')
).reset_index()

# データの出力
OUTPUT_PATH = output_folder + "/view_promotion_notification_output.csv"
df_view_promotion_notification.to_csv(OUTPUT_PATH, encoding='utf-8', index=False)

# compare_product, estimate_cost, view_productについて

- 入力したファイル一覧
    - compare_product,
    - estimate_cost,
    - view_product_detail_am,
    - view_product_gallery,
    - view_product_list,
    - view_product_color

def yymmdd_count(df, s):
    df_output = df.groupby(by='UID').agg(
        s = ('event_date', 'count')
    ).reset_index()
    return df_output

### 入力パス、出力パスの指定
PATH_DIR = input_tmp_folder
OUTPUT_PATH_DIR = output_folder

INPUT_FILENAME_LIST = [
    'compare_product_am',
    'estimate_cost',
    'view_product_detail_am',
    'view_product_gallery',
    'view_product_list',
    'view_product_color'
]

input_path_list = []
output_path_list = []

for name in INPUT_FILENAME_LIST:
    # input
    name_csv = name + '.csv'
    input_path = os.path.join(PATH_DIR, name_csv)
    input_path_list.append(input_path)

    # output
    output_name_csv = name_csv.replace(".csv", "_output.csv")
    output_path = os.path.join(OUTPUT_PATH_DIR, output_name_csv)
    output_path_list.append(output_path)

### 各ファイルについて処理
for input, output, filename in zip(input_path_list, output_path_list, INPUT_FILENAME_LIST):
    df_input = pd.read_csv(input, encoding="utf-8")
    df_output = yymmdd_count(df_input, filename)
    df_output.to_csv(output, encoding="utf-8", index=False)

# Change columns name of some files

# Change column names in file book_new_name.csv
book_new_name_cols = {'UID': 'UID',
'UID_counts_book': 'UID_counts_book',
'UID有無': 'UID有無',
'emp_id_count': 'app_mente_SA_select',
'emp_id_有無': 'SA_select',
'booking_date_count': 'app_reserve_mean'}


book_new_name_file = output_folder + "/book_new_name.csv"
pd.read_csv(book_new_name_file, encoding='utf-8').rename(columns=book_new_name_cols, errors='ignore').to_csv(book_new_name_file, index=False)
                                                                  
# Change column names in file DCSI_output_ver2.csv
DCSI_output_ver2_cols ={'UID': 'UID',
'Month': 'Month',
'Year': 'Year',
'Dealer': 'Dealer',
'Region': 'Region',
'REMINDER_': 'satisfaction_reminder',
'RECEPTION': 'satisfaction_reception',
'CUSTOMER_LOUNGE': 'satisfaction_customer_lounge',
'DELIVERY': 'satisfaction_delivery',
'REPAIR_QUALITY': 'satisfaction_repair_quality',
'FACILITY': 'satisfaction_facility',
'Q30Y': 'one_time_repair',
'OVERALL_SATISFACTION': 'total_satisfaction',
'Free_comment': 'Free_comment',
'follow_call': 'follow_call'}



DCSI_output_ver2_file = output_folder + "/DCSI_output_ver2.csv"
pd.read_csv(DCSI_output_ver2_file, encoding='utf-8').rename(columns=DCSI_output_ver2_cols, errors='ignore').to_csv(DCSI_output_ver2_file, index=False)
         
# Change column names in file ew_output.csv
ew_output_cols = {'UID': 'UID',
'ew_sales': 'ew_sales',
'ew_all_car': 'ew_all_car',
'ew_nun': 'ew_num'}

ew_output_file = output_folder + "/ew_output.csv"
pd.read_csv(ew_output_file, encoding='utf-8').rename(columns=ew_output_cols, errors='ignore').to_csv(ew_output_file, index=False)

# Change column names in file MH_new_name.csv
MH_new_name_cols ={'UID': 'UID',
'UID_counts_MyHonda': 'MH_Active',
'rate_service_point': 'satisfaction_service',
'rate_mechanic_point': 'satisfaction_mechanic',
'rate_chatbot_point': 'satisfaction_chatbot',
'rate_chatbot_agent_point': 'satisfaction_chatbot_agent',
'MyHonda有無': 'MH_ID'}
MH_new_name_file = output_folder + "/MH_new_name.csv"
pd.read_csv(MH_new_name_file, encoding='utf-8').rename(columns=MH_new_name_cols, errors='ignore').to_csv(MH_new_name_file, index=False)


# Drop redundant column in file sale_uni_min.csv
#sale_uni_min_file = root_folder + "/data/input/sale_uni_min.csv"
#pd.read_csv(sale_uni_min_file, encoding='utf-8').drop('Unnamed: 0', errors='ignore').to_csv(sale_uni_min_file, index=False)
